将VFS的read write操作更改至返回读取/写入大小

+++ src/x86_64/fs/vfs.c
@@ -242,14 +242,14 @@ int vfs_mount(const char *src, vfs_node_t node) {
     return VFS_STATUS_FAILED;
 }
 
-int vfs_read(vfs_node_t file, void *addr, size_t offset, size_t size) {
+size_t vfs_read(vfs_node_t file, void *addr, size_t offset, size_t size) {
     if (file == NULL || addr == NULL) return VFS_STATUS_FAILED;
     do_update(file);
     if (file->type == file_dir) return VFS_STATUS_FAILED;
     return callbackof(file, read)(file->handle, addr, offset, size);
 }
 
-int vfs_write(vfs_node_t file, void *addr, size_t offset, size_t size) {
+size_t vfs_write(vfs_node_t file, void *addr, size_t offset, size_t size) {
     if (file == NULL || addr == NULL) return VFS_STATUS_FAILED;
     do_update(file);
     if (file->type == file_dir) return VFS_STATUS_FAILED;
